fix conflict

+++ Assets/Script/Interface/IDamageReciever.cs
@@ -3,7 +3,7 @@ using Unity.Netcode;
 public interface IDamageReciever
 {
     public GameObject GameObject { get; }
-    public int CurrentHealth { get; }
-    public int MaxHealth { get; }
-    public void TakeDamage(int damage);
+    public float CurrentHealth { get; }
+    public float MaxHealth { get; }
+    public void TakeDamage(float damage);
 }
